Attempting merge

+++ src/Engine/SurfaceSet.h
@@ -1,5 +1,5 @@
 /*
- * Copyright 2010-2013 OpenXcom Developers.
+ * Copyright 2010-2015 OpenXcom Developers.
  *
  * This file is part of OpenXcom.
  *
@@ -19,7 +19,6 @@
 #ifndef OPENXCOM_SURFACESET_H
 #define OPENXCOM_SURFACESET_H
 
-#include <vector>
 #include <map>
 #include <string>
 #include <SDL.h>
@@ -60,7 +59,7 @@ public:
 	/// Gets the height of all frames.
 	int getHeight() const;
 	/// Gets the total frames in the set.
-	int getTotalFrames() const;
+	size_t getTotalFrames() const;
 	/// Sets the surface set's palette.
 	void setPalette(SDL_Color *colors, int firstcolor = 0, int ncolors = 256);
 	std::map<int, Surface*> *getFrames();
